% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_grid.R
\name{iso_grid}
\alias{iso_grid}
\title{Create hexagonal grid}
\usage{
iso_grid(x, resolution = 9)
}
\arguments{
\item{x}{An sf object the convex hull of which will determine the hex grid}

\item{resolution}{The resolution of the hex grid}
}
\value{
An sf object
}
\description{
Create a hexagonal grid from an sf object
}
\examples{
x = net_oldenburg_raw
iso_grid(x)
}
