% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_osrm.R
\name{iso_osrm}
\alias{iso_osrm}
\title{Calculate voronoi-style polygons based on grid with OSRM routing duration}
\usage{
iso_osrm(
  x,
  points,
  measure = "duration",
  osrm.server = getOption("osrm.server"),
  osrm.profile = getOption("osrm.profile")
)
}
\arguments{
\item{x}{An sf object of a grid}

\item{points}{An sf object of points}

\item{measure}{passed on to osrm::osrmTable, can be either "duration" (minutes) or "distance" (meters)}

\item{osrm.server}{the base URL of the routing server (end with "/"!)}

\item{osrm.profile}{the routing profile to use}
}
\value{
An sf object of a grid, where column 'index_min' gives the index of the input point, for which the measure is minimal
}
\description{
Calculate voronoi-style polygons based on grid with OSRM routing duration
}
\note{
The function requires OSRM to be running, with the server being

defined by the setting for the osrm package,

i.e., getOption("osrm.server") and getOption("osrm.profile")
}
\examples{
points = points_oldenburg
x = sf::st_as_sf(iso_grid(net_oldenburg_raw, resolution = 10))
grid = isonoi::iso_osrm(x, points, measure = "duration",
osrm.server = "http://router.project-osrm.org/", osrm.profile = "car")
plot(grid)
}
