% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_grid.R
\name{iso_vgrid}
\alias{iso_vgrid}
\title{Calculate voronoi-style polygons based on grid with Euclidean distances}
\usage{
iso_vgrid(points, grid)
}
\arguments{
\item{points}{An sf object of points}

\item{grid}{An sf object of a grid}
}
\description{
Calculate voronoi-style polygons based on grid with Euclidean distances
}
\note{
The function groups by the first column in the points object which should be unique (e.g. OSM ID).
}
\examples{
points = points_oldenburg
x = net_oldenburg_raw
grid = iso_grid(x)
vgrid = iso_vgrid(points, grid)
}
