% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_grid.R
\name{iso_vnoi}
\alias{iso_vnoi}
\title{Generate voronoi polygons from a bounding box and points}
\usage{
iso_vnoi(points, poly)
}
\arguments{
\item{points}{An sf object of points}

\item{poly}{A polygon object}
}
\value{
An sf object of voronoi polygons
}
\description{
Generate voronoi polygons from a bounding box and points
}
\examples{
points = points_oldenburg
poly = sf::st_convex_hull(sf::st_union(net_oldenburg_raw))
nvoi = iso_vnoi(points, poly)
plot(nvoi)
plot(points, add = TRUE)
}
