% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_oldenburg_raw}
\alias{net_oldenburg_raw}
\title{Dataset representing the road network in Oldenburg, Germany}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1738 rows and 9 columns.
}
\usage{
net_oldenburg_raw
}
\description{
See info in the 'data-raw' folder to see how this dataset was created
}
\examples{
net_oldenburg_raw
}
\keyword{datasets}
