% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points_oldenburg}
\alias{points_oldenburg}
\title{Dataset representing pubs in Oldenburg, Germany}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
points_oldenburg
}
\description{
See info in the 'data-raw' folder to see how this dataset was created
}
\examples{
net_oldenburg_raw
}
\keyword{datasets}
