% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_grid.R
\name{iso_grid}
\alias{iso_grid}
\title{Create hexagonal grid}
\usage{
iso_grid(x, ...)
}
\arguments{
\item{x}{An sf object the convex hull of which will determine the hex grid}

\item{...}{Extra arguments passed to the grid generator (TBC)}
}
\description{
Create hexagonal grid
}
\examples{
x = net_oldenburg_raw
iso_grid(x)
}
